spell-checker gets a new file!
